Extrai métodos para módulo de parsers
