+++ datasets/tests/management/commands/test_citycouncil.py
@@ -1,6 +1,7 @@
 from datetime import date, datetime
 
 import pytest
+
 from datasets.management.commands._citycouncil import save_agenda, save_attendance_list
