+++ scraper/spiders/cityhall.py
@@ -2,16 +2,12 @@ import re
 from datetime import date, datetime, timedelta
 
 import scrapy
+
+from datasets.parsers import from_str_to_datetime
 from scraper.items import CityHallBidItem, CityHallContractItem, CityHallPaymentsItem
 
 from . import BaseSpider
-from .utils import (
-    extract_param,
-    from_str_to_datetime,
-    identify_contract_id,
-    is_url,
-    strip_accents,
-)
+from .utils import extract_param, identify_contract_id, is_url, strip_accents
 
 
 class BidsSpider(BaseSpider):
