+++ scraper/spiders/utils.py
@@ -1,8 +1,9 @@
 import logging
 import re
 import unicodedata
-from datetime import datetime
 from urllib.parse import parse_qs, urlparse
+
+from datasets.parsers import from_str_to_date
 
 DOMAIN_FORMAT = re.compile(
     r"(?:^(\w{1,255}):(.{1,255})@|^)"
@@ -36,24 +37,6 @@ def extract_param(url, param):
         return value[0]
     except KeyError:
         return
-
-
-def from_str_to_datetime(date_str, supported_formats=["%d/%m/%Y", "%d/%m/%y"]):
-    if date_str is None:
-        return
-    for supported_format in supported_formats:
-        try:
-            return datetime.strptime(date_str, supported_format)
-        except ValueError:
-            pass
-
-
-def from_str_to_date(date_str, supported_formats=["%d/%m/%Y", "%d/%m/%y"]):
-    if date_str is None:
-        return
-    datetime_obj = from_str_to_datetime(date_str, supported_formats)
-    if datetime_obj:
-        return datetime_obj.date()
 
 
 def months_and_years(start_date, end_date):
